BotGetTime: change return type from double to float
Matches Q3 source signature 'float BotGetTime(bot_match_t *)'.
MSVC6 was emitting fld QWORD [0.0] on the early-return path because
the function return type was double. With float return type, the
compiler emits fld DWORD [0.0f] matching the reference.

MATCH 403->404, ICM 36->35

Source-commit: 7b0432df5116bbe9c21bd602470c1c60e15be429

+++ botlib/botlib.c
@@ -653,7 +653,7 @@ void __cdecl BotCTFSeekGoals(bot_state_t *bs);
 BOOL TeamPlayIsOn();
 BOOL __cdecl BotGetItemTeamGoal(char *String1, bot_goal_t *goal);
 int __cdecl BotGetMessageTeamGoal(bot_state_t *bs, char *String2, bot_goal_t *goal);
-double __cdecl BotGetTime(bot_match_t *match);
+float __cdecl BotGetTime(bot_match_t *match);
 // int __cdecl FindClientByName (1-param idb decl): see 3-param definition at ~L20657
 int __cdecl BotGetPatrolWaypoints(bot_state_t *bs, bot_match_t *match);
 int __cdecl BotAddressedToBot(bot_state_t *bs, bot_match_t *match);
@@ -23221,9 +23221,9 @@ int __cdecl BotGetMessageTeamGoal(bot_state_t *bs, char *String2, bot_goal_t *go
 }
 
 //----- (100267E0) --------------------------------------------------------
-double __cdecl BotGetTime(bot_match_t *match)
+float __cdecl BotGetTime(bot_match_t *match)
 {
-  double v1; // st7
+  float v1; // st7
   float v3; // [esp+0h] [ebp-18Ch]
   char String[152]; // [esp+4h] [ebp-188h] BYREF
   /* IDA split bot_match_t (240 B) into `char Destination[4]` + a phantom
